requirements.txt
flask
torch
opencv-python-headless
pillow
numpy
pandas
requests